requirements.txt
altgraph==0.10.2    
bdist-mpkg==0.5.0     
bonjour-py==0.3       
certifi==2019.11.28
chardet==3.0.4  
click==7.1.1  
dnspython==1.16.0 
Flask==1.1.2  
Flask-PyMongo==2.3.0  
Flask-WTF==0.14.3 
idna==2.9 
itsdangerous==1.1.0  
Jinja2==2.11.1 
macholib==1.5.1  
MarkupSafe==1.1.1  
matplotlib==1.3.1  
modulegraph==0.10.4 
numpy==1.8.0rc1  
pip==19.0.3 
py2app== 0.7.3  
pymongo==3.10.1 
pyobjc-core==2.5.1  
pyobjc-framework-Accounts==2.5.1  
pyobjc-framework-AddressBook==2.5.1  
pyobjc-framework-AppleScriptKit==2.5.1  
pyobjc-framework-AppleScriptObjC== .5.1  
pyobjc-framework-Automator==2.5.1  
pyobjc-framework-CFNetwork==2.5.1  
pyobjc-framework-Cocoa==2.5.1  
pyobjc-framework-Collaboration==2.5.1  
pyobjc-framework-CoreData==2.5.1  
pyobjc-framework-CoreLocation== 2.5.1  
pyobjc-framework-CoreText==  2.5.1  
pyobjc-framework-DictionaryServices== 2.5.1  
pyobjc-framework-EventKit==  2.5.1  
pyobjc-framework-ExceptionHandling==  2.5.1  
pyobjc-framework-FSEvents==2.5.1  
pyobjc-framework-InputMethodKit==2.5.1  
pyobjc-framework-InstallerPlugins==2.5.1  
pyobjc-framework-InstantMessage==2.5.1  
pyobjc-framework-LatentSemanticMapping== 2.5.1  
pyobjc-framework-LaunchServices==2.5.1     
pyobjc-framework-Message==2.5.1     
pyobjc-framework-OpenDirectory==2.5.1     
pyobjc-framework-PreferencePanes==2.5.1     
pyobjc-framework-PubSub==2.5.1     
pyobjc-framework-QTKit== 2.5.1     
pyobjc-framework-Quartz==2.5.1     
pyobjc-framework-ScreenSaver==2.5.1     
pyobjc-framework-ScriptingBridge==2.5.1     
pyobjc-framework-SearchKit==2.5.1     
pyobjc-framework-ServiceManagement==2.5.1     
pyobjc-framework-Social==2.5.1     
pyobjc-framework-SyncServices==2.5.1     
pyobjc-framework-SystemConfiguration==2.5.1     
pyobjc-framework-WebKit==2.5.1     
pyOpenSSL== 0.13.1    
pyparsing== 2.0.1     
python-dateutil==1.5       
python-dotenv==0.11.0    
pytz==2013.7    
requests==  2.23.0    
scipy==0.13.0b1  
setuptools==18.5      
six==1.4.1     
typing==3.7.4.1   
urllib3==1.25.8    
vboxapi==1.0       
Werkzeug==1.0.1     
WTForms==2.2.1     
xattr==0.6.4  